Finished with character creation.
