+++ Avelraangame3/Controllers/HomeController.cs
@@ -1,17 +1,16 @@
 using Microsoft.AspNetCore.Mvc;
 
-namespace Avelraangame3.Controllers
-{
-    public class HomeController() : Controller
-    {
-        public IActionResult Index()
-        {
-            return View();
-        }
+namespace Avelraangame3.Controllers;
 
-        public IActionResult Privacy()
-        {
-            return View();
-        }
+public class HomeController() : Controller
+{
+    public IActionResult Index()
+    {
+        return View();
+    }
+
+    public IActionResult Privacy()
+    {
+        return View();
     }
 }
