+++ Avelraangame3/Views/Home/Index.cshtml
@@ -4,8 +4,6 @@
     ViewData["Title"] = "Home Page";
 }
 
-@* <div id="model" hidden value="@Model"></div> *@
-
 <div class="text-center">
     <h3>some text goes here</h3>
     <p class="primary-green my-1">Some text stuff</p>
