+++ Avelraangame3/Views/Shared/_Layout.cshtml
@@ -24,7 +24,7 @@
                             <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Index">Home</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link text-dark" asp-controller="Character" asp-action="Index">Character</a>
+                            <a class="nav-link text-dark" asp-controller="Character" asp-action="Upload">Character</a>
                         </li>
                         <li class="nav-item">
                             <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Privacy">Privacy</a>
