--- Avelraangame3/wwwroot/js/Home/homeIndex.js
@@ -1,9 +1,0 @@
-$(() => {
-    $('#button').click(() => {
-        const model = $('#model').value;
-
-        console.log(model)
-    });
-
-
-});
