--- Models/Characters/CharacterAssets.cs
@@ -1,22 +1,0 @@
-﻿namespace Models;
-
-public class CharacterAssets
-{
-    public int HitpointsBase { get; set; }
-    public int HitpointsActual { get; set; }
-
-    public int ManaBase { get; set; }
-    public int ManaActual { get; set; }
-
-    public int ActionsBase { get; set; }
-    public int ActionsActual { get; set; }
-
-    public int DefenseBase { get; set; }
-    public int DefenseActual { get; set; }
- 
-    public int ResistBase { get; set; }
-    public int ResistActual { get; set; }
-
-    public int ReflexBase { get; set; }
-    public int ReflexActual { get; set; }
-}
