--- Models/Characters/CharacterCrafts.cs
@@ -1,16 +1,0 @@
-﻿namespace Models;
-
-public class CharacterCrafts
-{
-    public int Combat { get; set; } // dodge included
-    public int Arcane { get; set; }
-    public int Alchemy { get; set; } // potions and wearables
-    public int Psionics { get; set; }
-    public int Hunting { get; set; } // hide included
-    public int Advocacy { get; set; }
-    public int Mercantile { get; set; }
-    public int Tactics { get; set; }
-    public int Travelling { get; set; }
-    public int Sailing { get; set; }
-    public int Medicine { get; set; }
-}
