--- Models/Characters/CharacterIdentity.cs
@@ -1,7 +1,0 @@
-﻿namespace Models;
-
-public class CharacterIdentity
-{
-    public Guid Id { get; set; }
-    public Guid SessionId { get; set; }
-}
