--- Models/Characters/CharacterStats.cs
@@ -1,9 +1,0 @@
-﻿namespace Models;
-
-public class CharacterStats
-{
-    public int Strength { get; set; }
-    public int Athletics { get; set; }
-    public int Willpower { get; set; }
-    public int Abstract { get; set; }
-}
