--- Models/Characters/CharacterTraits.cs
@@ -1,8 +1,0 @@
-﻿namespace Models;
-
-public class CharacterTraits
-{
-    public string Race { get; set; }
-    public string Culture { get; set; }
-    public string Spec { get; set; }
-}
