+++ Models/Characters/CreateCharacter.cs
@@ -3,6 +3,7 @@
 public class CreateCharacter
 {
     public string Name { get; set; }
+    public string Portrait { get; set; }
     public string Race { get; set; }
     public string Culture { get; set; }
     public string Spec { get; set; }
