+++ Models/Characters/ImportCharacter.cs
@@ -4,3 +4,9 @@ public class ImportCharacter
 {
     public string CharacterString { get; set; }
 }
+
+public class ImportCharacterResponse
+{
+    public Guid CharacterId { get; set; }
+    public Guid SessionId { get; set; }
+}
